moved test to _MyViewTest class and delete ViewTest.java
